refactor(web): extract the Channels surface into domains/channels (LUM-2745)
The Channels tab grew inside domains/contacts before it became its own
top-level tab, leaving the contacts domain owning a surface it no longer
renders. Move the channel surface wholesale into a new domains/channels,
honoring the no-cross-domain-imports rule without growing the allowlist:

- domains/channels now owns channels-page, the Slack channel surface
  (list/section/card/override panel/tier legend/connection card/thread
  behavior), the adapter sub-tab list, and their hooks
  (use-channel-permission-overrides, use-channel-trust-floors,
  use-setup-channel-param) and queries.
- Code used by both domains hoists to the shared top level:
  assistant-display-name -> src/utils, use-assistant-channels and
  use-invite-link-dialog -> src/hooks, the invite-link dialog and share
  button -> src/components, and the shared channel UI types
  (ChannelStatus, AssistantChannelState) -> src/types/channel-types.
- src/utils/slack-workspace-cache.ts composes both domains' Slack query
  keys so disconnect/reconnect invalidation needs no cross-domain reach.
- The mixed detail+list test file splits along the same seam: the
  contact-card tests stay in contacts, the list-surface tests move next
  to AssistantChannelsList in channels.
- Moved stories retitle from Contacts/ to Channels/.

Pure reorganization: no behavior, copy, or styling changes.
