+++ clients/web/src/channels-page-route.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router";
 
 import { useActiveAssistantId } from "@/assistant/use-active-assistant-id";
-import { ChannelsPage } from "@/domains/contacts/channels-page";
+import { ChannelsPage } from "@/domains/channels/channels-page";
 import { navigateToNewConversation } from "@/utils/conversation-navigation";
 
 export function ChannelsPageRoute() {
