+++ clients/web/src/domains/channels/components/slack-channel-card.tsx
