+++ clients/web/src/domains/channels/components/slack-connection-card.tsx
