+++ clients/web/src/domains/channels/components/slack-thread-behavior.tsx
