+++ clients/web/src/domains/contacts/components/guardian-detail-view.tsx
@@ -6,7 +6,7 @@ import { Input } from "@vellumai/design-library/components/input";
 import { DetailCard } from "@/components/detail-card";
 import { ContactChannelsSection } from "@/domains/contacts/components/contact-channels-section";
 import { ContactTypeBadge } from "@/domains/contacts/components/contact-type-badge";
-import { ShareConnectionLinkButton } from "@/domains/contacts/components/share-connection-link-button";
+import { ShareConnectionLinkButton } from "@/components/share-connection-link-button";
 import type { ChannelInfo, ContactPayload } from "@/domains/contacts/types";
 
 interface GuardianDetailViewProps {
