+++ clients/web/src/domains/contacts/contacts-page.tsx
@@ -12,7 +12,7 @@ import { AssistantChannelsDetail } from "@/domains/contacts/components/assistant
 import { ContactDetailView } from "@/domains/contacts/components/contact-detail-view";
 import { ContactMergeDialog } from "@/domains/contacts/components/contact-merge-dialog";
 import { ContactsList } from "@/domains/contacts/components/contacts-list";
-import { GenerateInviteLinkDialog } from "@/domains/contacts/components/generate-invite-link-dialog";
+import { GenerateInviteLinkDialog } from "@/components/generate-invite-link-dialog";
 import { GuardianDetailView } from "@/domains/contacts/components/guardian-detail-view";
 import { LinkAccountDialog } from "@/domains/contacts/components/link-account-dialog";
 import { slackRosterOptions } from "@/domains/contacts/slack-users-query";
@@ -38,10 +38,10 @@ import {
 } from "@/generated/daemon/@tanstack/react-query.gen";
 import { channelsAvailableGet } from "@/generated/daemon/sdk.gen";
 import type { ChannelsAvailableGetResponse } from "@/generated/daemon/types.gen";
-import { assistantDisplayName } from "@/domains/contacts/assistant-display-name";
-import { useAssistantChannels } from "@/domains/contacts/hooks/use-assistant-channels";
+import { assistantDisplayName } from "@/utils/assistant-display-name";
+import { useAssistantChannels } from "@/hooks/use-assistant-channels";
 import { useChannelProvenance } from "@/domains/contacts/hooks/use-channel-provenance";
-import { useInviteLinkDialog } from "@/domains/contacts/hooks/use-invite-link-dialog";
+import { useInviteLinkDialog } from "@/hooks/use-invite-link-dialog";
 import { useAccountLink } from "@/domains/contacts/hooks/use-account-link";
 import { useAssistantFeatureFlagStore } from "@/stores/assistant-feature-flag-store";
 import { useAssistantIdentityStore } from "@/stores/assistant-identity-store";
