+++ clients/web/src/domains/contacts/types.ts
@@ -1,13 +1,16 @@
 import type {
   ChannelsAvailableGetResponse,
-  ChannelsReadinessGetResponse,
   ContactsGetResponse,
-  SlackChannelsGetResponse,
 } from "@/generated/daemon/types.gen";
-import type { SetupChannelId } from "@/types/channel-types";
 
 // Re-export shared channel types for domain consumers.
-export { SETUP_CHANNEL_IDS, isSetupChannelId, type SetupChannelId } from "@/types/channel-types";
+export {
+  SETUP_CHANNEL_IDS,
+  isSetupChannelId,
+  type AssistantChannelState,
+  type ChannelStatus,
+  type SetupChannelId,
+} from "@/types/channel-types";
 
 // ---------------------------------------------------------------------------
 // Types derived from the generated daemon SDK
@@ -17,12 +20,6 @@ export type ContactPayload = ContactsGetResponse["contacts"][number];
 export type ContactChannelPayload = ContactPayload["channels"][number];
 
 export type ChannelInfo = ChannelsAvailableGetResponse["channels"][number];
-
-export type SlackChannel = SlackChannelsGetResponse["channels"][number];
-
-type ReadinessSnapshot = ChannelsReadinessGetResponse["snapshots"][number];
-export type ChannelReadinessSnapshot = ReadinessSnapshot;
-export type ReadinessCheck = NonNullable<ReadinessSnapshot["localChecks"]>[number];
 
 // ---------------------------------------------------------------------------
 // UI-only types (no daemon/gateway equivalent)
@@ -38,13 +35,4 @@ export interface ContactSummary {
   role: string;
   contactType?: string | null;
   channelTypes?: string[];
-}
-
-export type ChannelStatus = "ready" | "incomplete" | "not_configured";
-
-export interface AssistantChannelState {
-  key: SetupChannelId;
-  status: ChannelStatus;
-  address?: string;
-  warning?: string;
 }
