+++ clients/web/src/types/channel-types.ts
@@ -12,3 +12,17 @@ export type SetupChannelId = (typeof SETUP_CHANNEL_IDS)[number];
 export function isSetupChannelId(value: string): value is SetupChannelId {
   return (SETUP_CHANNEL_IDS as readonly string[]).includes(value);
 }
+
+// ---------------------------------------------------------------------------
+// Channel setup state (UI-only; shared by the Channels tab and the Contacts
+// assistant detail)
+// ---------------------------------------------------------------------------
+
+export type ChannelStatus = "ready" | "incomplete" | "not_configured";
+
+export interface AssistantChannelState {
+  key: SetupChannelId;
+  status: ChannelStatus;
+  address?: string;
+  warning?: string;
+}
