+++ clients/web/src/utils/assistant-display-name.ts
